requirements.txt
beautifulsoup4==4.12.2
easyocr==1.7.1
lxml==4.9.3
Pillow==10.1.0
requests==2.31.0
selenium==4.15.2
torch==2.1.1
